# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend written in Go for the S2 ESGI Challenge",
        "title": "Backend",
        "contact": {}
    },
    "basePath": "/api",
    "paths": {
...
        "/schedules": {
            "get": {
                "description": "Get all schedule",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Get all schedule",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_esgi-challenge_backend_internal_models.Schedule"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "create new schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Create new schedule",
                "parameters": [
                    {
                        "description": "Schedule infos",
                        "name": "schedule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_esgi-challenge_backend_internal_models.ScheduleCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_esgi-challenge_backend_internal_models.Schedule"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/schedules/{id}": {
            "get": {
                "description": "Get schedule by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Get schedule by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_esgi-challenge_backend_internal_models.Schedule"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Update schedule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Schedule infos",
                        "name": "schedule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_esgi-challenge_backend_internal_models.ScheduleUpdate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_esgi-challenge_backend_internal_models.Schedule"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete schedule by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Delete schedule by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_esgi-challenge_backend_internal_models.Schedule": {
            "type": "object",
            "properties": {
                "campusId": {
                    "type": "integer"
                },
                "classId": {
                    "type": "integer"
                },
                "courseId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_esgi-challenge_backend_internal_models.ScheduleCreate": {
            "type": "object",
            "required": [
                "campusId",
                "classId",
                "courseId",
                "time"
            ],
            "properties": {
                "campusId": {
                    "type": "integer"
                },
                "classId": {
                    "type": "integer"
                },
                "courseId": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "github_com_esgi-challenge_backend_internal_models.ScheduleUpdate": {
            "type": "object",
            "required": [
                "campusId",
                "classId",
                "courseId",
                "time"
            ],
            "properties": {
                "campusId": {
                    "type": "integer"
                },
                "classId": {
                    "type": "integer"
                },
                "courseId": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}